+++ script/readme.rb
@@ -19,7 +19,6 @@ Say.call do
         say("Debug details about this ...", :debug)
         say("Info about stuff ...", :info)
         say("Maybe look into this thing ...", :warn)
-        say("Maybe look into the above thing ...", :warning)
         say("Failed to do a thing ...", :error)
 
         "The Result!"
@@ -44,7 +43,6 @@ Say.call do
         Say.("Debug details about this ...", :debug)
         Say.("Info about stuff ...", :info)
         Say.("Maybe look into this thing ...", :warn)
-        Say.("Maybe look into the above thing ...", :warning)
         Say.("Failed to do a thing ...", :error)
 
         "The Result!"